# [decoder] switching decoding library with safe casting