--- a/GLOCKFILE
+++ b/GLOCKFILE
@@ -4,5 +4,5 @@ github.com/cihub/seelog d2c6e5aa9fbfdd1c624e140287063c7730654115
 golang.org/x/tools 344080534465c5613c3563c30fd38c031073841d
 github.com/golang/lint 32a87160691b3c96046c0c678fe57c5bef761456
 github.com/stretchr/testify 8ce79b9f0b77745113f82c17d0756771456ccbd3
-github.com/ugorji/go faddd6128c66c4708f45fdc007f575f75e592a3c
+github.com/tinylib/msgp 362bfb3384d53ae4d5dd745983a4d70b6d23628c
 gopkg.in/ini.v1 12f418cc7edc5a618a51407b7ac1f1f512139df3