--- a/Rakefile
+++ b/Rakefile
@@ -21,7 +21,11 @@ PACKAGES = %w(
   ./statsd
 )
 
-EXCLUDE_LINT = 'span_gen.go'
+EXCLUDE_LINT = [
+    'model/services_gen.go',
+    'model/trace_gen.go',
+    'model/span_gen.go',
+]
 
 task :default => [:ci]
 
@@ -54,7 +58,11 @@ task :lint do
   error = false
   PACKAGES.each do |pkg|
     puts "golint #{pkg}"
-    output = `golint #{pkg} | grep -v #{EXCLUDE_LINT}`.strip
+    output = `golint #{pkg}`.split("\n")
+    output = output.reject do |line|
+      filename = line.split(':')[0]
+      EXCLUDE_LINT.include?(filename)
+    end
     if output.length > 0
       puts output
       error = true