--- a/model/compat.go
+++ b/model/compat.go
@@ -4,8 +4,8 @@ package model
 // grouping them by trace IDs
 // FIXME[1.x] this can be removed as we get pre-assembled traces from
 // clients
-func TracesFromSpans(spans []Span) []Trace {
-	var traces []Trace
+func TracesFromSpans(spans []Span) Traces {
+	traces := Traces{}
 	byID := make(map[uint64][]Span)
 	for _, s := range spans {
 		byID[s.TraceID] = append(byID[s.TraceID], s)