--- a/model/services.go
+++ b/model/services.go
@@ -7,6 +7,8 @@ import (
 	"reflect"
 )
 
+//go:generate msgp
+
 // ServicesMetadata is a standard key/val meta map attached to each named service
 type ServicesMetadata map[string]map[string]string