… agent/receiver_test.go
… model/client_api_test.go
… model/services_gen_test.go
… model/span_gen_test.go
… model/trace_gen.go
… model/trace_gen_test.go